[REFACTOR] 매물 DTO 패키지 이동
