+++ src/main/java/org/hanihome/hanihomebe/property/web/dto/response/basic/SharePropertyResponseDTO.java
@@ -1,4 +1,4 @@
-package org.hanihome.hanihomebe.property.web.dto.response;
+package org.hanihome.hanihomebe.property.web.dto.response.basic;
 
 import org.hanihome.hanihomebe.interest.region.Region;
 import org.hanihome.hanihomebe.item.web.dto.OptionItemResponseDTO;
@@ -31,7 +31,6 @@ public record SharePropertyResponseDTO(
         CostDetails costDetails,
         LivingConditions livingConditions,
         MoveInInfo moveInInfo,
-        ParkingOption parkingOption,
         String description,
         ShareInternalDetails internalDetails,                       // 2-6. 해당 매물의 층수
         CapacityShare capacityShare                             // 3. 수용 인원
@@ -81,7 +80,6 @@ public record SharePropertyResponseDTO(
                 shareProperty.getCostDetails(),
                 shareProperty.getLivingConditions(),
                 shareProperty.getMoveInInfo(),
-                shareProperty.getParkingOption(),
                 shareProperty.getDescription(),
                 // share 고유정보
                 shareProperty.getShareInternalDetails(),
